STYLE: Fix "sign-compare" warning in vtkMRMLColorLogic
This commit fixes a warning introduced by r24742 (ENH: Add terminology
information to color picker in Editor)


/path/to/Slicer/Libs/MRML/Logic/vtkMRMLColorLogic.cxx:101:45: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for (int i = 0; i < this->ColorFiles.size(); i++)
                                             ^
/path/to/Projects/Slicer/Libs/MRML/Logic/vtkMRMLColorLogic.cxx:106:49: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for (int i = 0; i < this->UserColorFiles.size(); i++)
                                                 ^
/path/to/Projects/Slicer/Libs/MRML/Logic/vtkMRMLColorLogic.cxx:112:56: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for (int i = 0; i < this->TerminologyColorFiles.size(); i++)
                                                        ^

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=24744
git-svn-id: http://svn.slicer.org/Slicer4/trunk@24744 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Logic/vtkMRMLColorLogic.cxx
@@ -98,18 +98,18 @@ void vtkMRMLColorLogic::PrintSelf(ostream& os, vtkIndent indent)
 
   os << indent << "UserColorFilePaths: " << this->GetUserColorFilePaths() << "\n";
   os << indent << "Color Files:\n";
-  for (int i = 0; i < this->ColorFiles.size(); i++)
+  for (size_t i = 0; i < this->ColorFiles.size(); i++)
     {
     os << indent.GetNextIndent() << i << " " << this->ColorFiles[i].c_str() << "\n";
     }
   os << indent << "User Color Files:\n";
-  for (int i = 0; i < this->UserColorFiles.size(); i++)
+  for (size_t i = 0; i < this->UserColorFiles.size(); i++)
     {
     os << indent.GetNextIndent() << i << " " << this->UserColorFiles[i].c_str() << "\n";
     }
 
   os << indent << "Terminology Color Files:\n";
-  for (int i = 0; i < this->TerminologyColorFiles.size(); i++)
+  for (size_t i = 0; i < this->TerminologyColorFiles.size(); i++)
     {
     os << indent.GetNextIndent() << i << " " << this->TerminologyColorFiles[i].c_str() << "\n";
     }
